Move ordinal stages into own directory
