+++ src/processTargets/modifiers/ordinalStages/AbsoluteOrdinalStage.ts
@@ -1,0 +1,19 @@
+import { Target } from "../../../typings/target.types";
+import { AbsoluteOrdinalScopeModifier } from "../../../typings/targetDescriptor.types";
+import { ProcessedTargetsContext } from "../../../typings/Types";
+import { ModifierStage } from "../../PipelineStages.types";
+import { createTarget, getTargets } from "./OrdinalStagesUtil";
+
+export class AbsoluteOrdinalStage implements ModifierStage {
+  constructor(private modifier: AbsoluteOrdinalScopeModifier) {}
+
+  run(context: ProcessedTargetsContext, target: Target): Target[] {
+    const targets = getTargets(context, target, this.modifier.scopeType);
+
+    const startIndex =
+      this.modifier.start + (this.modifier.start < 0 ? targets.length : 0);
+    const endIndex = startIndex + this.modifier.length - 1;
+
+    return [createTarget(target.isReversed, targets, startIndex, endIndex)];
+  }
+}
